reorganized posters and backgrounds
posters and backgrounds can be loaded automatically with the same name
as the collection in the assets folder (asset_folders: false).
This won't produce any errors and won't prevent the collection from
being created when a poster or background image isn't available through
file_poster or file_background.

This also makes adding/removing/changing the images easier because the
image itself has to be put into the assets folder and nothing else.