Updated Anime Collections to use MAL_IDs instead of the title
The Anime collections are now being organized through the MAL_IDs
because the title of the Anime could change based on the selected
language in the MyAnimeList Metadata Agent or when MyAnimeList changes
the title itself. The ID won't change and is more unique in that regard.

It does require the user to setup the MyAnimeList connection though.